requirements.txt
google-generativeai
instaloader
pillow
vinted_scraper
numpy
pandas